Free level Editor
I added the level editor that is supposedly part of the newer
platformer starter kit for XNA 4.0. I read of the terms and agreements
and it said that it was available for unrestricted use.
We are storing the levels in files in the project like the platformer
starter kit did , so It just took a little playing around with it to
figure out ho to get it to work with our tiles..
ADDED:
++Level Editor for Tile System
++New level 0 using the editor
Note -
You will need to change the content path in the level editor code to
match your repository. The editor was never designed with a local path
in mind, so until I get some more time to tweak it, you'll have to just
change the path name.