Enemy Class Abstraction Complete
I turned the original Enemy class into a parent class that the specific
enemies now inherit from. I left the handleCollisions() method and all
the stuff that is involved with loading or drawing the gun arm object or
bullets in the parent Enemy class.
DELETED:
--Redundant image file folder
FIXED:
=>Level Editor Drawing bug
- The level editor was treating all of the same letters as the same
object, which means it doesn't account for lower/upper case characters
being different. I changed the key to a different key and it resolved
the issue.
FOUND:
=> Level editor does not differentiate between upper and lowercase
chars, this could limit us in the future if we continue to expand our
resources.
NEXT TIME:
-Add new enemy types with some different AI code
-Text Helper Class tutorial I saw on the Web
-Vertical Moving Platforms
-Water Tiles
-New Sprites