Running Man AI
This fearless enemy runs blindly off cliffs and turns around at walls so
he can keep running .

INVESTIGATION REQUIRED: I need to explore restricting the updating of
these guys, or all of the enemies in general when they are not within
the outer viewport buffer for the game window. I've done something
similar for the drawing aspect, so it should be possible, and it will
likely be more efficient.