Vertical Moving Tiles (Bugged),  New Sprites, Ninja Class
I figured it would be simple to add the vertical block movement into the
game by using the horizontal  moving ones as an example.

While the tiles were platforms, the down movement of the player and
vertical movement block worked as expected.
When going up the platforms would not allow the player to ride on them
anymore, but it almost seemed like it would catch the player for a split
second while attempting to jump on the upward moving platforms.

I changed them to impassable tiles and the player can ride the block up,
but the player constantly attempts to fall into the block and is pushed
back up by collision detection. This would completely prevent the player
from jumping if the dbl jump was not in place.

The game currently considers a player to be on ground if the player's
bottom boundary is equal to the top boundary of the block, which means
that the player is not being moved vertically at a rate that is
consistent with the movement of the block. This results in the player
hovering or falling while on the platform. I will post an issue as a
reminder

ADDED:
=> Ninja Enemy class,