Initial SuperBeast AI Complete
If it sees you it gets mad and tries to trample or ram you to death.  It
uses the pacing AI to turn around and continue its rampage. I was
thinking about adding in  some kind of cooldown for it being mad, but
I'll address that at another time.

Need to change the pacing AI to use the current animation on turn
around.