5 bullet spreadshot implementation
I was able to get the collision with the enemies to work(there was an
error on my side). Currently when you fire, the spread shot is set as
default. We are able to switch to 3 bullets if necessary. The next
thing to be done is to go back to the regular shooting and only use the
spreadshot when the character goes through a powerup.